requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
docling==2.57.0
docling-core==2.48.4
docling-ibm-models==3.9.1
docling-parse==4.5.0
et_xmlfile==2.0.0
Faker==37.11.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.9.0
git-filter-repo==2.47.0
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2mathml==3.78.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
lxml==5.4.0
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.2
multiprocess==0.70.18
networkx==3.5
numba==0.62.1
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
polyfactory==2.22.2
pooch==1.8.2
psutil==7.1.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rtree==1.4.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.9
